requirements.txt
flask
mysql-connector-python
requests
sqlalchemy
jinja2
flask-cors